// src/api/openapi/model/stock.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderableType } from './orderableType';
import { Lock } from './lock';


export interface Stock { 
    type: OrderableType;
    name: any | null;
    id: any | null;
    displayable_name: any | null;
    lock: Lock;
}
export namespace Stock {
}